Factorize configuration-loading steps.
This exercise inadvertently added support for leading comments in the
current configuration system.  This is due to how the configuration
file is read (heuristically):

exec("self.global_config=" + file.read()) # Before
self.global_config = eval(file.read())    # After

In this commit, the leading comments are removed first by the 'eval'
function, and then the assignment is made to the variable.