Insulate multiprocessing test from parent environment.
When using environment variables in unit tests, there can be awkward
interactions between processes if 'forking' is used.  This commit,
which will be necessary for a forthcoming change, sets the starting
method to 'spawn', insulating the test from undesirable environment
collisions.

This commit also corrects a spelling error in the filename.