Adding log rotation by time
Changes to add ability to rotate the log files according to a time interval. The choice for whether to rotate log files according to time or file size is made in the config file (decision_engine.jsonnet) in the "file_rotate_by" parameter. The config file includes parameters for the unit of time and the time interval. The time interval is the same for both the DE log file and the channel log files, just as the size is.

RB : https://fermicloud140.fnal.gov/reviews/r/212/