Added --no-webserver to invocations of DEServer
In CI, all the unit tests would pass, but then the job to run the tests would time out.  This is because the CherryPy webserver was started up alongside, and never was shut down.  This would cause the job to eventually time out.  Passing this argument should ensure that this doesn't happen.