Add Countdown class to address global timeout problem.
The DE can be configured to shutdown hard after a specified time
period has elapsed (default is 10 seconds).  Unfortunately, this
timeout was applied for *each* channel and source process, resulting
in a wait of '(# channels) * 10 + (# sources) * 10' seconds.

This commit introduces a Countdown class which keeps track of the
elapsed time, enabling a global timeout (10 seconds, by default) that
is close to what the operator would expect.